requirements.txt
flask==2.3.3
python-telegram-bot==13.15
pandas==2.1.3
scikit-learn==1.3.2
xgboost==1.7.6
numpy==1.24.4
ta==0.10.2
schedule==1.2.1
joblib==1.3.2
requests==2.31.0